// back/src/generated/graphql/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * Le scalaire `DateTime` accepte des chaines de caractères
   * formattées selon le standard ISO 8601. Exemples:
   * - "yyyy-MM-dd" (eg. 2020-11-23)
   * - "yyyy-MM-ddTHH:mm:ss" (eg. 2020-11-23T13:34:55)
   * - "yyyy-MM-ddTHH:mm:ssX" (eg. 2020-11-23T13:34:55Z)
   * - "yyyy-MM-dd'T'HH:mm:ss.SSS" (eg. 2020-11-23T13:34:55.987)
   * - "yyyy-MM-dd'T'HH:mm:ss.SSSX" (eg. 2020-11-23T13:34:55.987Z)
   */
  DateTime: Date;
  /** Chaîne de caractère au format URL, débutant par un protocole http(s). */
  URL: URL;
  JSON: any;
};
// ...
export type BrokerReceipt = {
  __typename?: "BrokerReceipt";
  id: Scalars["ID"];
  /** Numéro de récépissé courtier */
  receiptNumber: Scalars["String"];
  /** Limite de validité du récépissé */
  validityLimit: Scalars["DateTime"];
  /** Département ayant enregistré la déclaration */
  department: Scalars["String"];
};
// ...
export type CompanySearchResult = {
  __typename?: "CompanySearchResult";
  /** SIRET de l'établissement */
  siret?: Maybe<Scalars["String"]>;
  /** État administratif de l'établissement. A = Actif, F = Fermé */
  etatAdministratif?: Maybe<Scalars["String"]>;
  /** Adresse de l'établissement */
  address?: Maybe<Scalars["String"]>;
  /** Code commune de l'établissement */
  codeCommune?: Maybe<Scalars["String"]>;
  /** Nom de l'établissement */
  name?: Maybe<Scalars["String"]>;
  /** Code NAF */
  naf?: Maybe<Scalars["String"]>;
  /** Libellé NAF */
  libelleNaf?: Maybe<Scalars["String"]>;
  /**
   * Installation classée pour la protection de l'environnement (ICPE)
   * associé à cet établissement
   */
  installation?: Maybe<Installation>;
  /** Récépissé transporteur associé à cet établissement (le cas échéant) */
  transporterReceipt?: Maybe<TransporterReceipt>;
  /** Récépissé négociant associé à cet établissement (le cas échant) */
  traderReceipt?: Maybe<TraderReceipt>;
  /** Récépissé courtier associé à cet établissement (le cas échant) */
  brokerReceipt?: Maybe<BrokerReceipt>;
  /** Agrément VHU démolisseur (le cas échéant, pour les profils VHU) */
  vhuAgrementDemolisseur?: Maybe<VhuAgrement>;
  /** Agrément VHU broyeur (le cas échéant, pour les profils VHU) */
  vhuAgrementBroyeur?: Maybe<VhuAgrement>;
};
// ...
export type Declaration = {
  __typename?: "Declaration";
  /** Année de la déclaration */
  annee?: Maybe<Scalars["String"]>;
  /** Code du déchet */
  codeDechet?: Maybe<Scalars["String"]>;
  /** Description du déchet */
  libDechet?: Maybe<Scalars["String"]>;
  /** Type de déclaration GEREP: producteur ou traiteur */
  gerepType?: Maybe<GerepType>;
};
// ...
export type Installation = {
  __typename?: "Installation";
  /** Identifiant S3IC */
  codeS3ic?: Maybe<Scalars["String"]>;
  /** URL de la fiche ICPE sur Géorisques */
  urlFiche?: Maybe<Scalars["String"]>;
  /** Liste des rubriques associées */
  rubriques?: Maybe<Array<Rubrique>>;
  /** Liste des déclarations GEREP */
  declarations?: Maybe<Array<Declaration>>;
};
// ...
export type Rubrique = {
  __typename?: "Rubrique";
  /**
   * Numéro de rubrique tel que défini dans la nomenclature des ICPE
   * Ex: 2710
   */
  rubrique: Scalars["String"];
  /** Alinéa pour la rubrique concerné */
  alinea?: Maybe<Scalars["String"]>;
  /** État de l'activité, ex: 'En fonct', 'À l'arrêt' */
  etatActivite?: Maybe<Scalars["String"]>;
  /** Régime autorisé pour la rubrique: déclaratif, autorisation, seveso, etc */
  regimeAutorise?: Maybe<Scalars["String"]>;
  /**
   * Description de l'activité:
   * Ex: traitement thermique de déchets dangereux
   */
  activite?: Maybe<Scalars["String"]>;
  /** Catégorie d'établissement associé: TTR, VHU, Traitement */
  category: Scalars["String"];
  /** Volume autorisé */
  volume?: Maybe<Scalars["String"]>;
  /** Unité utilisé pour le volume autorisé */
  unite?: Maybe<Scalars["String"]>;
  /** Type de déchets autorisé */
  wasteType?: Maybe<WasteType>;
};
// ...
export type TraderReceipt = {
  __typename?: "TraderReceipt";
  id: Scalars["ID"];
  /** Numéro de récépissé négociant */
  receiptNumber: Scalars["String"];
  /** Limite de validité du récépissé */
  validityLimit: Scalars["DateTime"];
  /** Département ayant enregistré la déclaration */
  department: Scalars["String"];
};
// ...
export type TransporterReceipt = {
  __typename?: "TransporterReceipt";
  id: Scalars["ID"];
  /** Numéro de récépissé transporteur */
  receiptNumber: Scalars["String"];
  /** Limite de validité du récépissé */
  validityLimit: Scalars["DateTime"];
  /** Département ayant enregistré la déclaration */
  department: Scalars["String"];
};
// ...
export type VhuAgrement = {
  __typename?: "VhuAgrement";
  id: Scalars["ID"];
  /** Numéro d'agrément VHU */
  agrementNumber: Scalars["String"];
  /** Département ayant enregistré la déclaration */
  department: Scalars["String"];
};
// ...
export type StitchingResolver<TResult, TParent, TContext, TArgs> =
  | LegacyStitchingResolver<TResult, TParent, TContext, TArgs>
  | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (
  obj: T,
  context: TContext,
  info: GraphQLResolveInfo
) => boolean | Promise<boolean>;
// ...
export type CompanySearchResultResolvers<
  ContextType = GraphQLContext,
  ParentType extends ResolversParentTypes["CompanySearchResult"] = ResolversParentTypes["CompanySearchResult"]
> = {
  siret?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  etatAdministratif?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  address?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  codeCommune?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  name?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  naf?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  libelleNaf?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  installation?: Resolver<
    Maybe<ResolversTypes["Installation"]>,
    ParentType,
    ContextType
  >;
  transporterReceipt?: Resolver<
    Maybe<ResolversTypes["TransporterReceipt"]>,
    ParentType,
    ContextType
  >;
  traderReceipt?: Resolver<
    Maybe<ResolversTypes["TraderReceipt"]>,
    ParentType,
    ContextType
  >;
  brokerReceipt?: Resolver<
    Maybe<ResolversTypes["BrokerReceipt"]>,
    ParentType,
    ContextType
  >;
  vhuAgrementDemolisseur?: Resolver<
    Maybe<ResolversTypes["VhuAgrement"]>,
    ParentType,
    ContextType
  >;
  vhuAgrementBroyeur?: Resolver<
    Maybe<ResolversTypes["VhuAgrement"]>,
    ParentType,
    ContextType
  >;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};